% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox.R
\name{create_sandbox}
\alias{create_sandbox}
\title{Create a sandbox environment as instructed.}
\usage{
create_sandbox(box = TRUE)
}
\arguments{
\item{sandbox}{either 1) TRUE in which case the returned environment is
is an environment (sandbox) with a parent below the global environment at
the time the call is made.  This is important because packages
loaded after the sandbox is created will not be visible to the
sandbox! 2) FALSE in which case the sandbox has as a parent whatever
environment is the parent.frame of the environment where this
function was called; or 3) an environment in which case the
returned sandbox has the environment as a parent.}
}
\value{
a new environment with the specified parent.
}
\description{
Create a sandbox environment as instructed.
}

