% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter-points.R
\name{fixed_parameter_points}
\alias{fixed_parameter_points}
\title{Expand a fixed grid of points in parameter space.}
\usage{
fixed_parameter_points(arg_values)
}
\arguments{
\item{arg_bounds, }{n_arg_points.}
}
\value{
data frame of points uniformly distributed in parameter
        space.
}
\description{
Expand a fixed grid of points in parameter space.
}

