% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter-points.R
\name{grid_parameter_points}
\alias{grid_parameter_points}
\title{Create a grid of points in parameter space within parameter-specific
bounds.}
\usage{
grid_parameter_points(arg_bounds, n_points)
}
\arguments{
\item{arg_bounds, }{n_arg_points.}
}
\value{
data frame of points uniformly distributed in parameter
        space.
}
\description{
Create a grid of points in parameter space within parameter-specific
bounds.
}

