% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter-trees.R
\name{is_leaf}
\alias{is_leaf}
\title{is_leaf}
\usage{
is_leaf(x)
}
\arguments{
\item{x}{potential leaf.}
}
\value{
boolean
}
\description{
For a list check whether a given entry is a leaf (in the parameter
definition language).
}
\details{
Parameter leaf nodes are defined by having a range (min and max).
}

