% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{move}
\alias{move}
\title{Function for moving a variable from one environment to another
by name.}
\usage{
move(x, from, to, clobber = FALSE)
}
\arguments{
\item{x}{name of variable to be moved.}

\item{from}{source environment}

\item{to}{target environment}

\item{clobber}{logical environment inidcating whether to allow
overwriting of variables in the target environment.}
}
\description{
Function for moving a variable from one environment to another
by name.
}

