% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{move_all_functions}
\alias{move_all_functions}
\title{Move function objects from one environment to another}
\usage{
move_all_functions(from, to)
}
\arguments{
\item{from}{environment to check for function objects. This
environment must be _only_ contain functions.}

\item{to}{environment to write funciton objects to.}
}
\description{
Move function objects from one environment to another
}

