% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter-points.R
\name{random_parameter_trees}
\alias{random_parameter_trees}
\title{Create a grid of points in parameter space with uniformly drawn
random marginals for each parameter within parameter-specific bounds.}
\usage{
random_parameter_trees(x, n_points)
}
\arguments{
\item{arg_tree, }{n_arg_points.}
}
\value{
data frame of points uniformly distributed in parameter
        space.
}
\description{
Create a grid of points in parameter space with uniformly drawn
random marginals for each parameter within parameter-specific bounds.
}

