% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicates.R
\name{replicates}
\alias{replicates}
\title{Function for running replicates of simulation.}
\usage{
replicates(functions = new.env(), data = new.env(),
  data_transformations = quote({ }), simulation = quote({ }),
  generated_quantities = quote({ }), parameters = list(),
  control = list(), sandbox = TRUE)
}
\arguments{
\item{functions}{An environment (could be coerced from a list) with
any additional functions needed by the simulation.}

\item{data}{An environment (could be coerced from a list) with all
immutable data for the simulation.  It may include code
snippets (e.g.-quote({a <- 3; o <- f(x)}))}

\item{data_transformations}{code snippet run once, created by quoting a
bracketed expression, (e.g.-quote({a <- 3; b <- 3*a})), which
will be used to transform data once prior to simulation.}

\item{simulation}{function defining a single simulation, taking
'parameters' as arguments.}

\item{parameters}{list of lists.  Each element is a list of parameter
values..}

\item{control}{control list for generating replicates, includes
elements: 1) size, the number of replicates to generate per 
parameter set; 2) output, a snippet run at the end of each
simulation, should persist results. 
used to summarize and persist simulation results.}

\item{sandbox}{if TRUE the parent.env where the call is made will not
be available to the code.  For full effects see
?create_sandbox.}

\item{simulation}{A code snippet, run once per replicate, created by 
quoting a bracketed expression, (e.g.-quote({a <- 3; b <-
-1*a}), which will be used to generate simulation output.  
This code can not modify the data.}
}
\description{
Function for running replicates of simulation.
}

