% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_one.R
\name{run_one}
\alias{run_one}
\title{Function for running a structured simulation.}
\usage{
run_one(functions = new.env(), data = new.env(),
  data_transformations = quote({ }), initial_states = new.env(),
  simulation = quote({ }), generated_quantities = quote({ }), n_steps = 1)
}
\arguments{
\item{functions}{An enviornment (could be coerced from a list) with
any additional functions needed by the simulation.}

\item{data}{An environment (could be coerced from a list) with all
immutable data for the simulation.}

\item{data_transformations}{code snippet run once, created by quoting a
bracketed expression, (e.g.-quote({a <- 3; b <- 3*a})), which
will be used to transform data once prior to simulation.}

\item{initial_states}{An environment (could be coerced from a list)
with the initial state of simulation parameters.}

\item{simulation}{A code snippet, run once per step, created by 
quoting a bracketed expression, (e.g.-quote({a <- 3; b <-
-1*a}), which will be used to transform the simulation state
once per step.  This code can not modify the data.}

\item{generated_quantities}{A code snippet, run once per step,
after simulation, which will be used to generate output 
dependent on the simulated state.}

\item{n_steps}{total number of simulation iterations to run.}
}
\description{
Function for running a structured simulation.
}

