% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{store}
\alias{store}
\title{Function for storing repeated values of a variable from 
one environment in another by name.}
\usage{
store(x, from, to, step, n_steps)
}
\arguments{
\item{x}{name of variable to be stored.}

\item{from}{source environment}

\item{to}{target environment}

\item{step}{simulation iteration}

\item{n_steps}{total number of simulation steps}
}
\description{
Function for storing repeated values of a variable from 
one environment in another by name.
}

