% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{store_all}
\alias{store_all}
\title{Function for storing repeated values of a variable from 
one environment.}
\usage{
store_all(from, to, step, n_steps)
}
\arguments{
\item{from}{source environment}

\item{to}{target environment}

\item{step}{simulation iteration}

\item{n_steps}{total number of simulation steps}
}
\description{
Function for storing repeated values of a variable from 
one environment.
}

