% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-helpers.R
\name{thin_forest}
\alias{thin_forest}
\title{Delete trees which are identical from a forest (list of trees).}
\usage{
thin_forest(forest)
}
\arguments{
\item{forest}{a list of trees.}
}
\value{
a list of trees without duplicates.
}
\description{
Delete trees which are identical from a forest (list of trees).
}

