% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-helpers.R
\name{tree_apply}
\alias{tree_apply}
\title{Tree apply}
\usage{
tree_apply(tree, leaf_def, action_def)
}
\arguments{
\item{leaf_def}{function which takes a node and returns TRUE only if
the node is a leaf.}

\item{action_def}{function to run on each leaf node.}

\item{x}{the tree-shaped list}
}
\value{
a list matching the tree structure and each node processed by
        the provided function.
}
\description{
Applies a function to tree leaves only.
}
\details{
In this terminology a tree is a multi-level list, 
not necessarily balanced, and each tree branch ends
in a node which is defined _only_ by the `leaf_def` function.
At each leaf the `action_def` function is applied.
}

