% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-helpers.R
\name{tree_delete_branches}
\alias{tree_delete_branches}
\title{Tree delete branches.}
\usage{
tree_delete_branches(tree, names)
}
\arguments{
\item{tree}{to delete branches from.}

\item{names}{a character vectors of parameters to remove.}
}
\value{
a list of named parameters.
}
\description{
Tree delete branches.
}

