% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-helpers.R
\name{tree_null_trim}
\alias{tree_null_trim}
\title{Remove NULL leaves tree.}
\usage{
tree_null_trim(tree)
}
\arguments{
\item{tree}{tree to trim.}
}
\value{
parameter tree with NULL values removed.
}
\description{
Remove NULL leaves tree.
}

