% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter-points.R
\name{is_leaf}
\alias{is_leaf}
\title{For a list check whether a given entry is a leaf (in the parameter
definition language).}
\usage{
is_leaf(x)
}
\description{
For a list check whether a given entry is a leaf (in the parameter
definition language).
}

