% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter-points.R
\name{tree_apply}
\alias{tree_apply}
\title{Tree (list) apply}
\usage{
tree_apply(x, leaf_def, action_def)
}
\arguments{
\item{x}{the tree-shaped list}

\item{leaf_def}{function which takes a node and returns TRUE only if
the node is a leaf.}

\item{action_def}{function to run on each leaf node.}
}
\value{
a list matching the tree structure and each node processed by
        the provided function.
}
\description{
Tree (list) apply
}

